# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.errorResponse:
    properties:
      error:
        example: invalid_credentials
        type: string
    type: object
...
  handler.transferRequest:
    properties:
      amount:
        example: 200
        type: number
      date:
        example: "2025-05-25"
        type: string
      description:
        example: Transferencia mensual
        type: string
      fee:
        example: 0
        type: number
      from_account_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      to_account_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  handler.transferResponse:
    properties:
      amount:
        example: 200
        type: number
      created_at:
        type: string
      date:
        type: string
      description:
        example: Transferencia mensual
        type: string
      fee:
        example: 0
        type: number
      from_account_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      from_account_name:
        example: Cuenta corriente
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      to_account_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      to_account_name:
        example: Ahorros
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST para gestión de gastos personales
  title: Expense Tracker API
  version: "1.0"
paths:
...
  /transfers:
    get:
      description: Devuelve las transferencias del usuario en un rango de fechas
      parameters:
      - description: Fecha inicio (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Fecha fin (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.transferResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Listar transferencias
      tags:
      - Transfers
    post:
      consumes:
      - application/json
      description: Transfiere fondos entre dos cuentas del usuario. Descuenta de la
        cuenta origen e incrementa la cuenta destino.
      parameters:
      - description: Datos de la transferencia
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.transferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.transferResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Crear transferencia
      tags:
      - Transfers
  /transfers/{id}:
    delete:
      description: Elimina una transferencia y revierte los saldos de ambas cuentas
      parameters:
      - description: ID de la transferencia
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar transferencia
      tags:
      - Transfers
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"